+++ RAPTOR/src/guidance/Pathfinder.h
@@ -7,6 +7,8 @@
 #define Pathfinder_h
 
 #include <math.h>
+#include "Arduino.h"
+
 #define pi 3.141592
 
 typedef struct Coordinate{
@@ -18,12 +20,12 @@ typedef struct Path{
     Coordinate lat_initial, long_initial;
     Coordinate lat_final, long_final;
     float lat_vec, long_vec;
-    double angle;
+    float angle;
 }Path;
 
 typedef struct PathAdjustment{
     double degrees; //Amount to turn
-    int rotation;   //Which way to turn. 1 is right, 0 is left.
+    uint8_t rotation;   //Which way to turn. 1 is right, 0 is left.
 }PathAdjustment;
 
 class Pathfinder{
